requirements.txt
requests
beautifulsoup4
pandas
pyyaml
streamlit
plotly
